fix: require whole-share quantities on buy and sell
quantity was typed float, so the API accepted sub-share positions the UI can
never represent; the frontend already enforces whole numbers and
trading_service already annotates quantity as int. Aligns the request schema
with both. Exchange amounts stay float — fractional currency is legitimate.

Addresses SECURITY_AUDIT.md Finding 6 (input-validation half).

+++ backend/app/schemas.py
@@ -3,12 +3,12 @@ from pydantic import BaseModel, Field
 
 class BuyRequest(BaseModel):
     ticker: str
-    quantity: float = Field(gt=0)
+    quantity: int = Field(gt=0)
 
 
 class SellRequest(BaseModel):
     ticker: str
-    quantity: float = Field(gt=0)
+    quantity: int = Field(gt=0)
 
 
 class ExchangeRequest(BaseModel):
